chore: repo's folder structure
