+++ src/Materials/Materials.hpp
@@ -1,0 +1,10 @@
+/*
+** EPITECH PROJECT, 2024
+** Raytracer
+** File description:
+** Materials
+*/
+
+#pragma once
+
+#include "Flat.hpp"
