+++ src/Raytracer.hpp
@@ -7,26 +7,18 @@
 
 #pragma once
 
-#include "RGBA.hpp"
-#include "Material.hpp"
 #include "Matrix/Matrix.hpp"
 
-#include <cmath>
-#include <iostream>
-#include <ostream>
-#include <vector>
-
 namespace RayTracer {
-    class Ray {
-    public:
-
-        Ray() = default;
-        Ray(const Point3D &origin, const Vector3D &direction)
-            : m_origin(origin)
-            , m_direction(direction)
-        {
-        }
-        Point3D m_origin;
-        Vector3D m_direction;
-    };
-}
+class Ray {
+public:
+    Ray() = default;
+    Ray(const Point3D &origin, const Vector3D &direction)
+        : m_origin(origin)
+        , m_direction(direction)
+    {
+    }
+    Point3D m_origin;
+    Vector3D m_direction;
+};
+} // namespace RayTracer
