+++ src/Shapes/Shapes.hpp
@@ -1,0 +1,10 @@
+/*
+** EPITECH PROJECT, 2024
+** Raytracer
+** File description:
+** Shapes
+*/
+
+#pragma once
+
+#include "Sphere.hpp"
