+++ src/main.cpp
@@ -5,13 +5,13 @@
 ** main
 */
 
-#include "Raytracer.hpp"
-#include "Scene.hpp"
 #include "Camera.hpp"
-#include "IShape.hpp"
+#include "Scene.hpp"
+
+#include "Materials/Materials.hpp"
+#include "Shapes/Shapes.hpp"
 
 #include <iostream>
-#include "Matrix/Matrix.hpp"
 
 int main()
 {
@@ -21,12 +21,12 @@ int main()
     RayTracer::Scene scene(800, 600);
 
     scene.setCamera(RayTracer::Camera());
-    scene.addShape(std::make_unique<RayTracer::Sphere>
-        (Point3D{0.1, 0, -1}, 0.5, yellow));
-    scene.addShape(std::make_unique<RayTracer::Sphere>
-        (Point3D{0, 0, -1}, 0.5, red));
-    scene.addShape(std::make_unique<RayTracer::Sphere>
-        (Point3D{1, 0, -1.5}, 0.5, blue));
+    scene.addShape(
+        std::make_unique<RayTracer::Sphere>(Point3D{0.1, 0, -1}, 0.5, yellow));
+    scene.addShape(
+        std::make_unique<RayTracer::Sphere>(Point3D{0, 0, -1}, 0.5, red));
+    scene.addShape(
+        std::make_unique<RayTracer::Sphere>(Point3D{1, 0, -1.5}, 0.5, blue));
     scene.generateImage(std::cout);
     return 0;
 }
